extract dragged item target
